requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
supabase==2.5.0
google-generativeai==0.5.4
firebase-admin==6.5.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
